BE use module structure
